Added a fallback to file creation date if date/time original doesn't exist
Caught a problem that if a file doesn't have the date/time original metadata then it returns a blank name.
I thought that making a warning and using the creation date would be best.
I suspect this will only happen with pictures that weren't taken with a digital camera (i.e. downloaded, screenshots, and possibly scans and pictures from undescript digital cameras)